requirements.txt
# DrugAge聊天机器人项目依赖

# 核心数据处理
pandas>=2.0.0
numpy>=1.24.0

# Web应用框架
streamlit>=1.28.0

# 数据可视化
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# AI和机器学习
openai>=1.3.0

# 统计分析
scipy>=1.11.0

# 环境变量管理
python-dotenv>=1.0.0

# 日志和工具
logging
pathlib
dataclasses
enum34;python_version<"3.4"

# 可选：增强功能
# scikit-learn>=1.3.0  # 如果需要机器学习功能
# nltk>=3.8  # 如果需要更高级的NLP功能
# spacy>=3.6  # 如果需要更复杂的实体识别

# 开发和测试
pytest>=7.4.0
pytest-cov>=4.1.0